remove close from txn
