+++ test/state/state.go
@@ -25,7 +25,7 @@ type State struct {
 	//
 	// This must be derived from the Rootstore.  For example it may be a namespace within the
 	// Rootstore, or even a transaction of that namespace.
-	Store corekv.Store
+	Store corekv.ReaderWriter
 
 	Txns []corekv.Txn
 }
